requirements.txt
# Top-level merged requirements for AssertMate
# Recommended Python: 3.10
# This file merges explicit pins found in subprojects (notably UTGen_LLM) and adds packages used by other components.
# Some packages are left unpinned where no explicit version was present in the repo (you can pin them if you need strict reproducibility).

annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
distro==1.9.0
dotmap
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jiter==0.8.0
loguru==0.7.2
openai==1.56.1
pydantic==2.10.3
pydantic_core==2.27.1
pyyaml==6.0.2
sniffio==1.3.1
tqdm==4.66.5
tree-sitter==0.23.2
tree-sitter-java==0.23.4
typing_extensions==4.12.2
